update host sever api.

+++ src/common/metadata/hostcontroller.go
@@ -87,10 +87,10 @@ type GetHostModuleIDsResult struct {
 }
 
 type ParamData struct {
-	ApplicationID       int   `json:"bk_biz_id"`
-	HostID              []int `json:"bk_host_id"`
-	OwnerModuleID       int   `json:"bk_owner_module_id"`
-	OwnerAppplicationID int   `json:"bk_owner_biz_id"`
+	ApplicationID       int64   `json:"bk_biz_id"`
+	HostID              []int64 `json:"bk_host_id"`
+	OwnerModuleID       int64   `json:"bk_owner_module_id"`
+	OwnerAppplicationID int64   `json:"bk_owner_biz_id"`
 }
 
 type AssignHostToAppParams struct {
@@ -107,9 +107,9 @@ type HostConfig struct {
 }
 
 type ModuleHostConfigParams struct {
-	ApplicationID int64 `json:"bk_biz_id"`
-	HostID        int64 `json:"bk_host_id"`
-	ModuleID      []int `json:"bk_module_id"`
+	ApplicationID int64   `json:"bk_biz_id"`
+	HostID        int64   `json:"bk_host_id"`
+	ModuleID      []int64 `json:"bk_module_id"`
 }
 
 type UserConfig struct {
